refactor Resource.tsx to allow filtering both topic type and resource type
previously, when filtering for a different topic, the resource type always reset to Videos. now changing one filter does not reset the other.
